[015] Migrate ArgumentException extensions into consolidated file
- Create ArgumentExceptionExtensions.cs with StringSegment and StringValues regions

- Change namespace from DotExtensions.MsExtensions.Exceptions to DotExtensions.MsExtensions.Primitives

- Delete legacy ArgumentException files and Exceptions directory

- Update GlobalUsings.cs to reference Primitives namespace instead of Exceptions

Refs: ticket-015, DECISIONS-DotExtensions-ms-extensions-collapse.md#D002
