+++ DotExtensions/MsExtensions/GlobalUsings.cs
@@ -4,5 +4,5 @@ global using System;
 global using System.Collections.Generic;
 global using System.Linq;
 global using System.Text;
-global using DotExtensions.MsExtensions.Exceptions;
+global using DotExtensions.MsExtensions.Primitives;
 global using Microsoft.Extensions.Primitives;
